move LocalizationController to another package
