+++ src/main/java/org/cmps/tetrahedron/interfaces/LocalizationListener.java
@@ -1,4 +1,4 @@
-package org.cmps.tetrahedron.enums;
+package org.cmps.tetrahedron.interfaces;
 
 import java.util.ResourceBundle;
